chore: add TopSectionImage widget separated and folder to AdvantagesSection
