+++ lib/app/pages/home/widgets/sections/advantages_section/advantages_section.dart
